refactor: move plugin setup to config.lua and update internal config usage
Internal functions now retrieve the config directly from config.lua
instead of receiving it via parameters. This change is internal and does
not affect the user API.
