+++ lua/undo-glow/types.lua
@@ -86,7 +86,6 @@
 ---@class UndoGlow.HandleHighlight : UndoGlow.RowCol
 ---@field bufnr integer Buffer number.
 ---@field ns? integer Namespace id.
----@field config UndoGlow.Config Configuration for undo-glow.
 ---@field state UndoGlow.State Current state of the highlight.
 
 ---Represents a region (row/column coordinates) in the buffer.
